+++ src/main/kotlin/image/ImageFilter.kt
@@ -1,0 +1,12 @@
+package org.example.image
+
+import org.example.kernels.Kernel
+import org.example.parallelismStrategies.ParallelismStrategy
+
+
+interface ImageFilter {
+    fun writeToFile()
+    fun applyParallelismStrategy(parallelismStrategy: ParallelismStrategy)
+    fun applyKernel(kernel: Kernel)
+}
+
